+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](https://semver.org).
 ### Added
 
 * Abbreviation for OpenBitTorrent tracker.
+* Abbreviation for Torrent Club tracker.
 
 ### Changed