+++ CHANGELOG.md
@@ -17,6 +17,10 @@ This project adheres to [Semantic Versioning](https://semver.org).
 
 * Neaten the files output with ``--show-files``.
 * Neaten the abbrs output.
+* Refactor ``open`` and ``random`` abbreviation implementation.
+	* Removed from constants.
+	* Implemented in ``replace_abbreviations`` function.
+	* This change allows multiple random trackers to be used in one call.
 
 ### Fixed